Redoing tweak on jumps since using classes got buggy

+++ src/CharactersServer/Characters/CharacterServer.h
@@ -182,7 +182,7 @@ private:
 
     virtual void moveRight(int distance, int vel, Box *boxOfEnemy, bool isGrounded) = 0;
 
-    virtual void moveLeft(int distance, int vel, Box *boxOfEnemy, bool isGrounded) = 0;
+    virtual void moveLeft(int distance, float vel, Box *boxOfEnemy, bool isGrounded) = 0;
 
     virtual void grip();
