+++ src/CharactersServer/Characters/RyuServer.h
@@ -21,7 +21,7 @@ public:
 
 private:
     void moveRight(int distance, int vel, Box *boxOfEnemy, bool isGrounded);
-    void moveLeft(int distance, int vel, Box *boxOfEnemy, bool isGrounded);
+    void moveLeft(int distance, float vel, Box *boxOfEnemy, bool isGrounded);
     void stand();
     void update(int distance, int posContrincante, actions_t actionRecieved, Box *boxEnemy) override;
