+++ src/CharactersServer/Characters/WolverineServer.h
@@ -15,7 +15,7 @@ public:
 
 private:
     void moveRight(int distance, int vel, Box *boxOfEnemy, bool isGrounded);
-    void moveLeft(int distance, int vel, Box *boxOfEnemy, bool isGrounded);
+    void moveLeft(int distance, float vel, Box *boxOfEnemy, bool isGrounded);
     void stand();
 
     int getSpriteNumber();
